MR: dxmin dxmax now global extrema

+++ src/io_mpi2.f90
@@ -301,7 +301,7 @@ module Io
       call MPI_TYPE_COMMIT (global_type, mpi_err)
       call check_success ('output', 'commit global type', file)
 !
-      call delete_file (trim (directory_snap)//'/'//file)
+      if (lroot) call delete_file (trim (directory_snap)//'/'//file)
       call MPI_FILE_OPEN (mpi_comm, trim (directory_snap)//'/'//file, MPI_MODE_CREATE+MPI_MODE_WRONLY, io_info, handle, mpi_err)
       call check_success ('output', 'open', trim (directory_snap)//'/'//file)
 !
